Makes borg reagent containers not be inserted into various machines.

+++ code/game/machinery/teleporter.dm
@@ -65,9 +65,10 @@
 
 
 		if(istype(L, /obj/effect/landmark/) && istype(L.loc, /turf))
+			if(!user.unEquip(I))
+				return
 			to_chat(usr, "You insert the coordinates into the machine.")
 			to_chat(usr, "A message flashes across the screen reminding the traveller that the nuclear authentication disk is to remain on the [station_name()] at all times.")
-			user.drop_item()
 			qdel(I)
 
 			if(C.data == "Clown Land")
